Use same batch size during evaluation and train
It seems that memory usage is about the same, and trying to use a
larger batch size crashes due to lack of memory in evaluation. I'm not
sure if this is a regression after switching to recurrent models, or if
this has always been the case and I didn't run into the issue due to
small batch sizes.