Call clearState before starting training
Calling

trainer:train()
evaluator:evaluate()
trainer:train()

in succession was causing a large blow up in memory at the second call
to train(). Adding this clearState call fixes this issue for the most
part; I believe the evaluate() might be storing some intermediate
activations that are kept around when model:train() is called?