Critical fix: Use lyaml instead of yaml
This is a critical fix because the yaml library does not follow some key
YAML specs. Specifically, the keyword "False" is interpreted as a
string, giving it a truthy value rather than a falsey one. This means
that every time a config used "False" as a keyword, it was actually
interpreted as "True"!.