Call _update after clearState
To be honest, I don't recall why I did this. I think it has something to
do with the fact that calling self:_update ensures that self.modules is
populated, so that calls to :type(), :training(), etc. will operate on
the modules. However, in practice, I've overridden :type and :training
to handle this (so that a module added after a call to :training or
:evaluate will still have the same mode as the rest of the modules in
the container). But, the code is here, and it's easier to commit than
test my experiments without it, so I'm committing as is.