Don't remove modules 1 and 2 from self.modules
We need self.modules to be populated with one copy of self.init and
self.residual for various reasons (e.g. so that `:parameters()` still
works correctly after a call to `:clearState()`). The issue with
`:parameters` was what caused the following issue, and this commit fixes
it: https://github.com/torch/cunn/issues/362