Do not call clearState on model while training
Calling clearState() seems to cause issues that, after 4-5 days of
debugging, I haven't been able to fix. See, for example:

https://github.com/torch/nn/issues/1141
https://github.com/torch/cunn/issues/441

Further, it's unclear to me if `getParameters` and memory management in
general works well when a call to `clearState` can destroy modules (and
therefore weight tensors). The easiest solution to all of this is simply
to never call clearState on the model while it is training.

When saving the model, we create a copy of it on the CPU, and call
clearState on this CPU copy, which we then save to disk.